fix different handling of whitespace in python 3.13 causing hashing inconsistencies
this is the culprit behind https://github.com/understanding-search/maze-dataset/issues/48

a note we added to `maze_dataset_config.py`:

this was causing hashing issues on 3.13 vs older versions because somehow,
the `__doc__` variable is different across versions??????? WHY???????? IT TREATS WHITESPACE DIFFERENTLY
so we just uh. strip it all now.

see:
https://github.com/understanding-search/maze-dataset/actions/runs/14028046497/job/39270080746?pr=53
https://github.com/understanding-search/maze-dataset/actions/runs/14028046497/job/39270080742?pr=53
https://www.diffchecker.com/tqIMSevy/